+++ CHANGELOG.md
@@ -16,6 +16,11 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - `Query::decoded` the string representation of the component decoded.
 - `URLSearchParams::decoded` the string representation of the component decoded.
 - `tryNew` named constructor added to all components class to returns a new instance on success or `null` on failure.
+- `Query::normalized`
+- `Path::nornalized`
+- `Fragment::normalized`
+- `HierarchicalPath::normalized`
+- `Datapath::normalized`
 
 ### Fixed